# docs/swagger.yaml
basePath: /
...
host: localhost:8080
info:
  contact: {}
  description: A messaging service that processes and sends messages through webhooks
    with scheduling capabilities.
  title: Insider Messaging API
  version: "1.0"
paths:
  /api/v1/messages/sent:
    get:
      consumes:
      - application/json
      description: Retrieves a list of sent messages with pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get sent messages
      tags:
      - messages
...
swagger: "2.0"